Split the unrelated es methods from the es_operations
